refactor(mcp): share the listing helpers the grants feature already has
The Activity listing landed in parallel with main's OAuth grants listing,
so both hand-rolled the same project filter, member lookup, project-name
lookup and privileged-user resolution, and shared carried two identical
platform-wide sentinels. One mcp-listing-utils module now backs both, the
Activity response is a plain SeekPage, and the one-method retention wrapper
folds into the service that already owns the table.
