+++ packages/server/api/src/app/file/file.module.ts
@@ -5,7 +5,7 @@ import { agentRetention } from '../ee/agent/agent-retention'
 import { SystemJobName } from '../helper/system-jobs/common'
 import { systemJobHandlers } from '../helper/system-jobs/job-handlers'
 import { systemJobsSchedule } from '../helper/system-jobs/system-job'
-import { mcpActivityRetention } from '../mcp/activity/mcp-activity-retention'
+import { mcpActivityService } from '../mcp/activity/mcp-activity-service'
 import { fileService } from './file.service'
 import { filesController, signedStepFileController } from './files-controller'
 
@@ -14,7 +14,7 @@ export const fileModule: FastifyPluginAsyncZod = async (app) => {
     systemJobHandlers.registerJobHandler(SystemJobName.FILE_CLEANUP_TRIGGER, async () => {
         await fileService(app.log).deleteStaleBulk([FileType.FLOW_RUN_LOG, FileType.FLOW_RUN_LOG_SLICE, FileType.FLOW_STEP_FILE, FileType.TRIGGER_EVENT_FILE, FileType.TRIGGER_PAYLOAD, FileType.WEBHOOK_PAYLOAD, FileType.MCP_CALL_PAYLOAD])
         await agentRetention(app.log).deleteStaleFlowStepConversations()
-        await mcpActivityRetention(app.log).deleteStale()
+        await mcpActivityService(app.log).deleteStale()
     })
     await systemJobsSchedule(app.log).upsertJob({
         job: {
