+++ packages/server/api/src/app/mcp/oauth/token/mcp-oauth-grants.controller.ts
@@ -5,12 +5,11 @@ import {
     PrincipalType,
     RevokeMcpOAuthGrantsRequestBody,
 } from '@activepieces/shared'
-import { FastifyRequest } from 'fastify'
 import { FastifyPluginAsyncZod } from 'fastify-type-provider-zod'
 import { StatusCodes } from 'http-status-codes'
 import { z } from 'zod'
 import { securityAccess } from '../../../core/security/authorization/fastify-security'
-import { userService } from '../../../user/user-service'
+import { mcpListingUtils } from '../../mcp-listing-utils'
 import { mcpOAuthTokenService } from './mcp-oauth-token.service'
 
 export const mcpOAuthGrantsController: FastifyPluginAsyncZod = async (app) => {
@@ -18,7 +17,7 @@ export const mcpOAuthGrantsController: FastifyPluginAsyncZod = async (app) => {
     app.get('/v1/mcp-oauth/grants', ListGrantsRequest, async (req): Promise<SeekPage<McpOAuthGrant>> => {
         return mcpOAuthTokenService.listGrants({
             platformId: req.principal.platform.id,
-            userId: await resolveUserIdFilter(req),
+            userId: await mcpListingUtils.resolveUserIdFilter(req),
             projectIds: req.query.projectIds,
             memberIds: req.query.memberIds,
             clientKeys: req.query.clientKeys,
@@ -31,15 +30,10 @@ export const mcpOAuthGrantsController: FastifyPluginAsyncZod = async (app) => {
         await mcpOAuthTokenService.revokeGrants({
             ids: req.body.ids,
             platformId: req.principal.platform.id,
-            userId: await resolveUserIdFilter(req),
+            userId: await mcpListingUtils.resolveUserIdFilter(req),
         })
         return reply.status(StatusCodes.NO_CONTENT).send()
     })
-}
-
-async function resolveUserIdFilter(req: FastifyRequest): Promise<string | null> {
-    const user = await userService(req.log).getOneOrFail({ id: req.principal.id })
-    return userService(req.log).isUserPrivileged(user) ? null : req.principal.id
 }
 
 const ListGrantsRequest = {
