+++ packages/server/api/src/app/mcp/oauth/token/mcp-oauth-token.service.ts
@@ -1,14 +1,13 @@
 import { randomBytes } from 'crypto'
 import { ActivepiecesError, apId, ErrorCode, isNil, sanitizeObjectForPostgresql, SeekPage, spreadIfDefined, unique } from '@activepieces/core-utils'
 import { cryptoUtils } from '@activepieces/server-utils'
-import { McpOAuthClientKey, McpOAuthGrant, McpOAuthToken, PLATFORM_WIDE_PROJECT_FILTER_VALUE, UserWithMetaInformation } from '@activepieces/shared'
-import { Brackets, In, ObjectLiteral, SelectQueryBuilder } from 'typeorm'
+import { McpOAuthClientKey, McpOAuthGrant, McpOAuthToken } from '@activepieces/shared'
+import { In, ObjectLiteral, SelectQueryBuilder } from 'typeorm'
 import { repoFactory } from '../../../core/db/repo-factory'
 import { JwtAudience, jwtUtils } from '../../../helper/jwt-utils'
 import { buildPaginator } from '../../../helper/pagination/build-paginator'
 import { paginationHelper } from '../../../helper/pagination/pagination-utils'
-import { projectRepo } from '../../../project/project-repo'
-import { mapToUserWithMetaInformation, userRepo } from '../../../user/user-service'
+import { mcpListingUtils } from '../../mcp-listing-utils'
 import { mcpOAuthClientIdentity } from '../client/mcp-oauth-client-identity'
 import { McpOAuthClientEntity } from '../client/mcp-oauth-client.entity'
 import { mcpOAuthPkce } from '../mcp-oauth.pkce'
@@ -164,14 +163,14 @@ export const mcpOAuthTokenService = {
         if (!isNil(memberIds)) {
             queryBuilder.andWhere(`${TOKEN_ALIAS}."userId" IN (:...memberIds)`, { memberIds })
         }
-        applyProjectFilter(queryBuilder, projectIds)
+        mcpListingUtils.applyProjectFilter({ queryBuilder, alias: TOKEN_ALIAS, projectIds })
 
         const { data, cursor: nextCursor } = await paginator.paginate(queryBuilder)
 
         const [clientNames, members, projectNames] = await Promise.all([
             findClientNames({ clientIds: data.filter((token) => isNil(token.clientKey) || token.clientKey === UNKNOWN_CLIENT_KEY).map((token) => token.clientId) }),
-            findMembers({ userIds: data.map((token) => token.userId), platformId }),
-            findProjectNames({ projectIds: data.map((token) => token.projectId), platformId }),
+            mcpListingUtils.findMembers({ userIds: data.map((token) => token.userId), platformId }),
+            mcpListingUtils.findProjectNames({ projectIds: data.map((token) => token.projectId), platformId }),
         ])
 
         const rows = data.map((token) => ({
@@ -214,22 +213,6 @@ function applyGrantScope<T extends ObjectLiteral>(queryBuilder: SelectQueryBuild
     }
 }
 
-function applyProjectFilter<T extends ObjectLiteral>(queryBuilder: SelectQueryBuilder<T>, projectIds: string[] | undefined): void {
-    if (isNil(projectIds)) {
-        return
-    }
-    const scopedProjectIds = projectIds.filter((projectId) => projectId !== PLATFORM_WIDE_PROJECT_FILTER_VALUE)
-    const includesPlatformWide = projectIds.length !== scopedProjectIds.length
-    queryBuilder.andWhere(new Brackets((qb) => {
-        if (scopedProjectIds.length > 0) {
-            qb.orWhere(`${TOKEN_ALIAS}."projectId" IN (:...scopedProjectIds)`, { scopedProjectIds })
-        }
-        if (includesPlatformWide) {
-            qb.orWhere(`${TOKEN_ALIAS}."projectId" IS NULL`)
-        }
-    }))
-}
-
 async function findClientNames({ clientIds }: FindClientNamesParams): Promise<Map<string, string | null>> {
     const distinct = unique(clientIds)
     if (distinct.length === 0) {
@@ -237,27 +220,6 @@ async function findClientNames({ clientIds }: FindClientNamesParams): Promise<Ma
     }
     const clients = await clientRepo().findBy({ clientId: In(distinct) })
     return new Map(clients.map((client) => [client.clientId, client.clientName]))
-}
-
-async function findProjectNames({ projectIds, platformId }: FindProjectNamesParams): Promise<Map<string, string>> {
-    const distinct = unique(projectIds.filter((projectId): projectId is string => !isNil(projectId)))
-    if (distinct.length === 0) {
-        return new Map()
-    }
-    const projects = await projectRepo().findBy({ id: In(distinct), platformId })
-    return new Map(projects.map((project) => [project.id, project.displayName]))
-}
-
-async function findMembers({ userIds, platformId }: FindMembersParams): Promise<Map<string, UserWithMetaInformation>> {
-    const distinct = unique(userIds)
-    if (distinct.length === 0) {
-        return new Map()
-    }
-    const users = await userRepo().find({ where: { id: In(distinct), platformId }, relations: { identity: true } })
-    return new Map(users.flatMap((user) => {
-        const member = mapToUserWithMetaInformation(user)
-        return isNil(member) ? [] : [[user.id, member] as const]
-    }))
 }
 
 export class OAuthTokenError extends Error {
@@ -297,16 +259,6 @@ type RevokeRefreshTokenParams = {
 
 type FindClientNamesParams = {
     clientIds: string[]
-}
-
-type FindProjectNamesParams = {
-    projectIds: (string | null)[]
-    platformId: string
-}
-
-type FindMembersParams = {
-    userIds: string[]
-    platformId: string
 }
 
 type GrantScope = {
