+++ packages/server/api/test/unit/app/mcp/mcp-activity-recorder.test.ts
@@ -26,7 +26,6 @@ const mcp: ProjectScopedMcpServer = {
     type: McpServerType.PROJECT,
     token: apId(),
     disabledTools: [],
-    flows: [],
 }
 
 describe('MCP activity recording predicate', () => {
